split models.stream in multiple files
