+++ tests/test_stream_app.py
@@ -4,9 +4,9 @@ from pytest_mock import MockerFixture
 
 from corva.application import Corva
 from corva.configuration import SETTINGS
-from corva.models.stream import (
-    BaseStreamContext,
-    LogType,
+from corva.models.stream.context import BaseStreamContext
+from corva.models.stream.log_type import LogType
+from corva.models.stream.raw_stream import (
     RawAppMetadata,
     RawDepthRecord,
     RawMetadata,
